requirements.txt
beautifulsoup4==4.13.4
discord.py==2.5.2
fastapi==0.116.1
pymongo==4.13.2
python-dotenv==1.1.1
Requests==2.32.4
uvicorn[standard]
aiohttp~=3.12.15